feat: make extension optional
